# apps/api/src/modules/tournaments/dto/tournament-created-by.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import type { TournamentCreatedByData } from '@repo/domain';

export class TournamentCreatedByDto implements TournamentCreatedByData {
  @ApiProperty()
  id!: string;

  @ApiProperty()
  login!: string;

  @ApiProperty()
  name!: string;
}